The DigitalDLSorterDNN Class

The DigitalDLSorterDNN object stores all the information related to Deep Neural Network model. It contains the trained model, the training history and the results of prediction on test data. After running calculateEvalMetrics, it is possible to find the performance evaluation of the model on test data (see ?calculateEvalMetrics for details).

Details

The steps related to Deep Learning are carried out using keras package which uses R6 classes. If you want to save the object as an RDS file, digitalDLSorteR provides a saveRDS generic function that transforms the model stored as an R6 object into a native valid R object. Specifically, the model is converted into a list with the architecture of the network and the weights learned during training. That is the minimum information needed to use the model as predictor. If you want to maintain the optimizer state, see ?saveTrainedModelAsH5. If you want to store DigitalDLSorter object as an RDA file, see ?preparingToSave.

Slots

model

Trained Deep Neural Network. This slot can contain a R6 keras.engine.sequential.Sequential object or a list with two elements: the architecture of the model and the resulting weights after training.

training.history

List with the evolution of selected metrics during training.

test.metrics

Performance of the model on test data.

test.pred

Deconvolution results on test data. Columns are cell types, rows are samples and each entry corresponds to the proportion of this cell type in this sample.

cell.types

Vector with cell types to deconvolute.

features

Vector with features used during training. These features will be used in subsequent predictions (the nomenclature used must be the same).

test.deconv.metrics

Model performance on each sample of test data compared to known cell proportions. This slot is used after ?calculateEvalMetrics, so see ?calculateEvalMetrics for more details.