Deconvolute bulk RNA-Seq samples using a pre-trained DigitalDLSorter model

Deconvolute bulk gene expression samples (bulk RNA-Seq) to enumerate and quantify the proportion of cell types present in a bulk sample using Deep Neural Network models. This function is intended for users who want to use pre-trained models integrated in the package. So far, the available models allow to deconvolute the immune infiltration of breast cancer (using data from Chung et al., 2017) and the immune infiltration of colorectal cancer (using data from Li et al., 2017) samples. For the former, two models are available at two different levels of specificity: specific cell types (breast.chung.specific) and generic cell types (breast.chung.generic). See breast.chung.generic, breast.chung.specific, and colorectal.li documentation from the digitalDLSorteRdata package for more details.

deconvDigitalDLSorter(
  data,
  model = NULL,
  batch.size = 128,
  normalize = TRUE,
  simplify.set = NULL,
  simplify.majority = NULL,
  verbose = TRUE
)

Arguments

data
Matrix or data frame with bulk RNA-Seq samples with genes as rows in SYMBOL notation and samples as columns.

model
Pre-trained DNN model to use to deconvolute data. Up to now, the available models are intended to deconvolute samples from breast cancer (breast.chung.generic and breast.chung.specific) and colorectal cancer (colorectal.li). These pre-trained models are stored in the digitalDLSorteRdata package, so it must be installed together with digitalDLSorteR to use this function.

batch.size
Number of samples loaded into RAM each time during the deconvolution process. If not specified, batch.size will be set to 128.

normalize
Normalize data before deconvolution (TRUE by default).

simplify.set
List specifying which cell types should be compressed into a new label whose name will be the list name item. See examples and vignettes for details.

simplify.majority
List specifying which cell types should be compressed into the cell type with the highest proportion in each sample. Unlike simplify.set, this argument allows to maintain the complexity of the results while compressing the information, as no new labels are created.

verbose
Show informative messages during execution.

Value

A data frame with samples (\(i\)) as rows and cell types (\(j\)) as columns. Each entry represents the predicted proportion of cell type \(j\) in sample \(i\).

Details

This function is intended for users who want to use digitalDLSorteR to deconvolute their bulk RNA-Seq samples using pre-trained models. For users who want to build their own models from other scRNA-Seq datasets, see the loadSCProfiles and deconvDigitalDLSorterObj functions.

References

Chung, W., Eum, H. H., Lee, H. O., Lee, K. M., Lee, H. B., Kim, K. T., et al. (2017). Single-cell RNA-seq enables comprehensive tumour and immune cell profiling in primary breast cancer. Nat. Commun. 8 (1), 15081. doi: doi: 10.1038/ncomms15081 .

See also

Examples

if (requireNamespace("digitalDLSorteRdata", quietly = TRUE)) {
  # to ensure compatibility
  tensorflow::tf$compat$v1$disable_eager_execution()
  library(digitalDLSorteRdata)
  data(breast.chung.specific)
  data(TCGA.breast.small)
  results1 <- deconvDigitalDLSorter(
    data = TCGA.breast.small,
    model = breast.chung.specific,
    normalize = TRUE
  )
  # simplify arguments
  simplify <- list(Tumor = c("ER+", "HER2+", "ER+/HER2+", "TNBC"),
                   Bcells = c("Bmem", "BGC"))
  # in this case the names of the list will be the new labels
  results2 <- deconvDigitalDLSorter(
    TCGA.breast.small,
    model = breast.chung.specific,
    normalize = TRUE,
    simplify.set = simplify
  )
  # in this case the cell type with the highest proportion will be the new label
  results3 <- deconvDigitalDLSorter(
    TCGA.breast.small,
    model = breast.chung.specific,
    normalize = TRUE,
    simplify.majority = simplify
  )
}
#> === Filtering 32876 features in data that are not present in trained model
#> === Setting 11305 features that are not present in trained model to zero
#> === Normalizing data
#> === Predicting cell types present in the provided samples
#> Warning: `predict_generator` is deprecated. Use `predict` instead, it now accept generators.
#> DONE
#> === Filtering 32876 features in data that are not present in trained model
#> === Setting 11305 features that are not present in trained model to zero
#> === Normalizing data
#> === Predicting cell types present in the provided samples
#> Warning: `predict_generator` is deprecated. Use `predict` instead, it now accept generators.
#> DONE
#> === Filtering 32876 features in data that are not present in trained model
#> === Setting 11305 features that are not present in trained model to zero
#> === Normalizing data
#> === Predicting cell types present in the provided samples
#> Warning: `predict_generator` is deprecated. Use `predict` instead, it now accept generators.
#> DONE