Install Python dependencies for digitalDLSorteR

This is a helper function to install Python dependencies needed: a Python interpreter with TensorFlow Python library and its dependencies. It is performed using the reticulate package and the installer of the tensorflow R package. The available options are virtual or conda environments. The new environment is called digitaldlsorter-env. In any case, this installation can be manually done as it is explained in https://diegommcc.github.io/digitalDLSorteR/articles/kerasIssues.html, but we recommend using this function.

installTFpython(
  conda = "auto",
  python.version = "3.8",
  tensorflow.version = "2.6",
  install.conda = FALSE,
  miniconda.path = NULL
)

Arguments

conda

Path to a conda executable. Use "auto" (by default) allows reticulate to automatically find an appropriate conda binary.

python.version

Python version to be installed in the environment ("3.8" by default). We recommend keeping this version as it has been tested to be compatible with tensorflow 2.6.

tensorflow.version

Tensorflow version to be installed in the environment ("2.6" by default).

install.conda

Boolean indicating if install miniconda automatically using reticulate. If TRUE, conda argument is ignored. FALSE by default.

miniconda.path

If install.conda is TRUE, you can set the path where miniconda will be installed. If NULL, conda will find automatically the proper place.

Value

No return value, called for side effects: installation of conda environment with a Python interpreter and Tensorflow

Details

This function is intended to make easier the installation of the requirements needed to use digitalDLSorteR. It will automatically install Miniconda (if wanted, see Parameters) and create an environment called 'digitaldlsorter-env'. If you want to use other python/conda environment, see ?tensorflow::use_condaenv and/or the vignettes.

Examples

if (FALSE) { # \dontrun{
notesInstallation <- installTFpython(
  method = "auto", conda = "auto", install.conda = TRUE
)
} # }