Save DigitalDLSorter objects as RDS files

Save DigitalDLSorter and DigitalDLSorterDNN objects as RDS files. keras models cannot be stored natively as R objects (e.g. RData or RDS files). By saving the structure as a JSON-like character object and the weights as a list, it is possible to retrieve the model and make predictions. If the trained.model slot is empty, the function will behave as usual. Note: with this option, the state of optimizer is not saved, only the architecture and weights. It is possible to save the entire model as an HDF5 file with the saveTrainedModelAsH5 function and to load it into a DigitalDLSorter object with the loadTrainedModelFromH5 function. See documentation for details.

saveRDS(
  object,
  file,
  ascii = FALSE,
  version = NULL,
  compress = TRUE,
  refhook = NULL
)

# S4 method for DigitalDLSorterDNN
saveRDS(
  object,
  file,
  ascii = FALSE,
  version = NULL,
  compress = TRUE,
  refhook = NULL
)

# S4 method for DigitalDLSorter
saveRDS(
  object,
  file,
  ascii = FALSE,
  version = NULL,
  compress = TRUE,
  refhook = NULL
)

Arguments

object

DigitalDLSorter or DigitalDLSorterDNN object to be saved

file

File path where the object will be saved

ascii

a logical. If TRUE or NA, an ASCII representation is written; otherwise (default), a binary one is used. See the comments in the help for save.

version

the workspace format version to use. NULL specifies the current default version (3). The only other supported value is 2, the default from R 1.4.0 to R 3.5.0.

compress

a logical specifying whether saving to a named file is to use "gzip" compression, or one of "gzip", "bzip2" or "xz" to indicate the type of compression to be used. Ignored if file is a connection.

refhook

a hook function for handling reference objects.

Value

No return value, saves a DigitalDLSorter object as an RDS file on disk.

See also