Show distribution plots of the cell proportions generated by generateBulkCellMatrix

Show distribution plots of the cell proportions generated by generateBulkCellMatrix. These frequencies will determine the proportion of different cell types used during the simulation of pseudo-bulk RNA-Seq samples. There are 6 subsets of proportions generated by different approaches that can be visualized in three ways: box plots, violin plots and lines plots. You can also plot the probabilities based on the number of different cell types present in the samples by setting type.plot = 'nCellTypes'.

showProbPlot(object, type.data, set, type.plot = "boxplot")

Arguments

object
DigitalDLSorter object with prob.cell.types slot with plot slot.

type.data
Subset of data to show: train or test.

set
Integer determining which of the 6 different subsets to display.

type.plot
Character determining which type of visualization to display. It can be 'boxplot', 'violinplot', 'linesplot' or 'ncelltypes'. See Description for more information.

Value

ggplot object.

Details

These plots are only for diagnostic purposes. This is the reason because they are generated without any parameter introduced by the user.

See also

Examples

# simulating data
sce <- SingleCellExperiment::SingleCellExperiment(
  matrix(
    rpois(100, lambda = 5), nrow = 40, ncol = 30, 
    dimnames = list(paste0("Gene", seq(40)), paste0("RHC", seq(30)))
  ),
  colData = data.frame(
    Cell_ID = paste0("RHC", seq(30)),
    Cell_Type = sample(x = paste0("CellType", seq(4)), size = 30, 
                       replace = TRUE)
  ),
  rowData = data.frame(
    Gene_ID = paste0("Gene", seq(40))
  )
)
DDLS <- loadSCProfiles(
  single.cell.data = sce,
  cell.ID.column = "Cell_ID",
  gene.ID.column = "Gene_ID"
)
probMatrix <- data.frame(
  Cell_Type = paste0("CellType", seq(4)),
  from = c(1, 1, 1, 30),
  to = c(15, 15, 50, 70)
)
DDLS <- generateBulkCellMatrix(
  object = DDLS,
  cell.ID.column = "Cell_ID",
  cell.type.column = "Cell_Type",
  prob.design = probMatrix,
  num.bulk.samples = 60
)
#> 
#> === The number of bulk RNA-Seq samples that will be generated is equal to 60
#> Error in generateBulkCellMatrix(object = DDLS, cell.ID.column = "Cell_ID",     cell.type.column = "Cell_Type", prob.design = probMatrix,     num.bulk.samples = 60) Not all cell types considered in DigitalDLSorter object are in test data. digitalDLSorteR needs to have all cell types in both subsets (training and test). Please, provide a bigger single-cell experiment or consider simulate new single-cell profiles with 'estimateZinbwaveParams' and 'simSCProfiles' functions
lapply(
  X = 1:6, FUN = function(x) {
      showProbPlot(
        DDLS,
        type.data = "train",
        set = x,
        type.plot = "boxplot"
      )
    }
  )
#> Error in showProbPlot(DDLS, type.data = "train", set = x, type.plot = "boxplot") 'prob.cell.types' slot is empty