Simulate new single-cell expression profiles using ZINB-WaVE model parameters

Simulate single-cell expression profiles by randomly sampling from a negative binomial distribution and inserting dropouts by sampling from a binomial distribution using ZINB -WaVE parameters estimated by estimateZinbwaveParams function.

simSCProfiles(
  object,
  cell.ID.column,
  cell.type.column,
  n.cells,
  suffix.names = "_Simul",
  cell.types = NULL,
  file.backend = NULL,
  name.dataset.backend = NULL,
  compression.level = NULL,
  block.processing = FALSE,
  block.size = 1000,
  chunk.dims = NULL,
  verbose = TRUE
)

Arguments

object
DigitalDLSorter object with single.cell.real and zinb.params slots.

cell.ID.column
Name or column number corresponding to cell names of expression matrix in cells metadata.

cell.type.column
Name or column number corresponding to the cell type of each cell in cells metadata.

n.cells
Number of simulated cells generated by cell type (i.e. if you have 10 different cell types in your dataset, if n.cells = 100, then 1000 cell profiles will be simulated).

suffix.names
Suffix used in simulated cells. This suffix must be unique in simulated cells, so ensure that this suffix does not appear in real cell names.

cell.types
Vector indicating which cell types you want to simulate. If NULL (by default), n.cells single-cell profiles for all cell types will be simulated.

file.backend
Valid file path to store simulated single-cell expression profiles as HDF5 file (NULL by default). If provided, data is stored in HDF5 files as back-end by using DelayedArray, HDF5Array and rhdf5 packages instead of loading in RAM memory. This is suitable for situations where you have a large amount of data that cannot be allocated in memory. Note that operations on this data will be carried out by blocks (i.e subsets of determined size) which can lead to longer execution times.

name.dataset.backend
Name of dataset in HDF5 file that will be used. Note that it cannot exist. If NULL (by default), a random dataset name will be used.

compression.level
The compression level used if file.backend is provided. It is an integer value between 0 (no compression) and 9 (highest and slowest compression). See ?getHDF5DumpCompressionLevel from HDF5Array package for more information.

block.processing
Boolean indicating if data should be simulated by blocks (only if file.backend is used, FALSE by default). This functionality is suitable for cases where is not possible to allocate data in memory and it leads to longer execution times.

block.size
Number of single-cell expression profiles that will be simulated in each iteration during the process. Higher numbers resulting in more memory usage but lesser execution times. Set according to available computational resources (1000 by default). Note that it cannot be greater than the total number of simulated cells.

chunk.dims
Specifies the dimensions that HDF5 chunk will have. If NULL, the default value is a vector of two items: the number of genes considered by ZINB-WaVE model during the simulation and only one sample in order to increase the read times in the following steps. Greater number of columns written in each chunk can lead to longer read times in subsequent steps. Note that it cannot be greater than dimensions of simulated matrix.

verbose
Show informative messages during the execution (TRUE by default).

Value

A DigitalDLSorter object with single.cell.simul slot containing a SingleCellExperiment object with the simulated single-cell expression profiles.

Details

Before this step, see ?estimateZinbwaveParams. As described in Torroja and Sanchez-Cabo, 2019, this function simulates a determined number of transcriptional profiles for each cell type provided by randomly sampling from a negative binomial distribution with \(\mu\) and \(\theta\) estimated parameters and inserting dropouts by sampling from a binomial distribution with pi probability. All parameters are estimated from single-cell real data using estimateZinbwaveParams function. It uses the ZINB-WaVE model (Risso et al., 2018). For more details about the model, see ?estimateZinbwaveParams and Risso et al., 2018.

file.backend argument allows to create a HDF5 file with simulated single-cell profiles that will be used as back-end in order to work with data allocated on disk instead of RAM memory. If you use file.backend argument with block.processing = FALSE, all single-cell profiles will be simulated in one step and, therefore, allocated in RAM memory. Then, data will be written in HDF5 file. In order to avoid collapsing RAM memory if so many single-cell profiles are simulated, single-cell profiles can be simulated and written in HDF5 files by blocks of block.size size by setting block.processing = TRUE.

References

Risso, D., Perraudeau, F., Gribkova, S. et al. (2018). A general and flexible method for signal extraction from single-cell RNA-seq data. Nat Commun 9, 284. doi: 10.1038/s41467-017-02554-5 Torroja, C. and Sánchez-Cabo, F. (2019). digitalDLSorter: A Deep Learning algorithm to quantify immune cell populations based on scRNA-Seq data. Frontiers in Genetics 10, 978. doi: 10.3389/fgene.2019.00978

See also

Examples

if (requireNamespace("digitalDLSorteRdata", quietly = TRUE)) {
  library(digitalDLSorteRdata)
  data(DDLSChung)
  DDLSChung <- simSCProfiles(
    object = DDLSChung,
    cell.ID.column = "Cell_ID",
    cell.type.column = "Cell_type",
    n.cells = 10,
    verbose = TRUE
  )
}
#> === Get parameters from model:
#>     - mu: 22, 18751
#>     - pi: 22, 18751
#>     - Theta: 18751
#> === Selected cell type(s) from ZINB-WaVE model (13 cell type(s)):
#>     - Bmem
#>     - DC
#>     - ER+
#>     - ER+ and HER2+
#>     - HER2+
#>     - Macrophage
#>     - Monocyte
#>     - Stromal
#>     - TCD8
#>     - Tme
#>     - TNBC
#>     - Treg
#>     - BGC
#> === Simulated matrix dimensions:
#>     - n (cells): 130
#>     - J (genes): 18751
#>     - i (# entries): 2437630
#> 
#> DONE