Deconvolute bulk gene expression samples (bulk RNA-Seq)

Deconvolute bulk gene expression samples (bulk RNA-Seq). This function requires a DigitalDLSorter object with a trained Deep Neural Network model (trained.model slot) and the new bulk RNA-Seq samples to be deconvoluted in the deconv.data slot. See ?loadDeconvData for more details.

deconvDigitalDLSorterObj(
  object,
  name.data,
  batch.size = 128,
  normalize = TRUE,
  simplify.set = NULL,
  simplify.majority = NULL,
  verbose = TRUE
)

Arguments

object
DigitalDLSorter object with trained.data and deconv.data slots.

name.data
Name of the data stored in the DigitalDLSorter object. If not provided, the first data set will be used.

batch.size
Number of samples per gradient update. If not specified, batch.size will default to 128.

normalize
Normalize data before deconvolution (TRUE by default).

simplify.set
List specifying which cell types should be compressed into a new label whose name will be the list item. See examples for details. If provided, results are stored in a list with 'raw' and 'simpli.set' results.

simplify.majority
List specifying which cell types should be compressed into the cell type with the highest proportion in each sample. Unlike simplify.set, it allows to maintain the complexity of the results while compressing the information, as no new labels are created. If provided, the results are stored in a list with 'raw' and 'simpli.majority' results.

verbose
Show informative messages during the execution.

Value

DigitalDLSorter object with deconv.results slot. The resulting information is a data frame with samples (\(i\)) as rows and cell types (\(j\)) as columns. Each entry represents the proportion of \(j\) cell type in \(i\) sample. If simplify.set or/and simpplify.majority are provided, the deconv.results slot will contain a list with raw and simplified results.

Details

This function is intended for users who have built a devonvolution model using their own single-cell RNA-Seq data. If you want to use a pre-trained model to deconvolute your samples, see ?deconvDigitalDLSorter.

References

Torroja, C. and Sánchez-Cabo, F. (2019). digitalDLSorter: A Deep Learning algorithm to quantify immune cell populations based on scRNA-Seq data. Frontiers in Genetics 10, 978. doi: doi: 10.3389/fgene.2019.00978

See also

Examples

if (FALSE) {
if (!requireNamespace("digitalDLSorteRdata", quietly = TRUE)) {
  install.packages(
    "digitalDLSorteRdata", 
    repos = "https://diegommcc.github.io/digitalDLSorteRdataRepo/"
  )
}
library(digitalDLSorteRdata)
data(DDLSLi.list)
DDLSLi <- listToDDLS(DDLSLi.list)
data(TCGA.colon.se)
# to ensure compatibility
tensorflow::tf$compat$v1$disable_eager_execution()
# simplify arguments
simplify = list(Macrophages = c("Mc", "M"))
DDLSLi <- loadDeconvData(
  object = DDLSLi, data = TCGA.colon.se,
  name.data = "TCGA.colon"
)
DDLSLi <- deconvDigitalDLSorterObj(
  object = DDLSLi,
  name.data = "TCGA.colon",
  simplify.set = simplify,
  simplify.majority = simplify
)
}